Fix sound side crashing "probably"
