+++ src/main/java/cassiokf/industrialrenewal/blocks/railroad/BlockCrossingRail.java
@@ -1,7 +1,7 @@
 package cassiokf.industrialrenewal.blocks.railroad;
 
 
-import cassiokf.industrialrenewal.IRSoundHandler;
+import cassiokf.industrialrenewal.init.IRSoundRegister;
 import net.minecraft.client.resources.I18n;
 import net.minecraft.client.util.ITooltipFlag;
 import net.minecraft.creativetab.CreativeTabs;
@@ -34,7 +34,7 @@ public class BlockCrossingRail extends BlockNormalRailBase {
         long thisTime = System.currentTimeMillis();
         if ((thisTime - lastTime) >= PERIOD) {
             lastTime = thisTime;
-            world.playSound(null, pos, IRSoundHandler.TILEENTITY_TRAINHORN, SoundCategory.BLOCKS, 2F, 1F);
+            world.playSound(null, pos, IRSoundRegister.TILEENTITY_TRAINHORN, SoundCategory.BLOCKS, 2F, 1F);
         }
     }
