+++ src/main/java/cassiokf/industrialrenewal/tileentity/tubes/TileEntityEnergyCableGauge.java
@@ -18,6 +18,7 @@ public abstract class TileEntityEnergyCableGauge extends TileEntityEnergyCable
         IBlockState state = world.getBlockState(pos);
         facing = state.getBlock() instanceof BlockEnergyCableGauge
                 ? state.getValue(BlockEnergyCableGauge.FACING) : EnumFacing.NORTH;
+        Sync();
         return facing;
     }
 
@@ -31,8 +32,8 @@ public abstract class TileEntityEnergyCableGauge extends TileEntityEnergyCable
     {
         int outputs = getMaster().getOutPutCount();
         float currentAmount = (float) getMaster().getOutPut() / (outputs > 0 ? (float) outputs : 1f);
-        float totalCapacity = (float) energyContainer.getMaxOutput();
-        currentAmount = currentAmount / totalCapacity;
+        float totalCapacity = (float) getMaster().energyContainer.getMaxOutput();
+        currentAmount = Utils.normalize(currentAmount, 0, totalCapacity);
         amount = Utils.lerp(amount, currentAmount, 0.1f);
         return amount * 90f;
     }
